Update Dart compiler and regenerate machine outputs

+++ tests/machine/x/dart/map_nested_assign.dart
@@ -3,6 +3,6 @@ var data = {
 };
 
 void main() {
-  (data['outer'] as Map)['inner'] = 2;
+  (data['outer'] as Map)['inner'] = (2 as int);
   print(((data as Map)['outer'] as Map)['inner']);
 }
