+++ tests/machine/x/dart/var_assignment.dart
@@ -1,6 +1,6 @@
 var x = 1;
 
 void main() {
-  x = 2;
+  x = (2 as int);
   print(x);
 }
